Refactor: Modularize codebase and add default symbol fetching
I've refactored the Binance data downloader into a more modular structure:
- Core data structures moved to `core.py`.
- Downloader logic separated into `downloader.py`.
- URL building logic moved to `url_builder.py`.
- Utility functions (logging, argument parsing) moved to `utils.py`.
- `main.py` updated to orchestrate the components.
- Tests refactored to mirror the new module structure.

New Feature: Default to all trading symbols if none are specified.
- If the `--symbols` argument is not provided, the application will now fetch
  all symbols with a "TRADING" status from the relevant Binance APIs
  (Spot, Futures UM/CM, Options).
- Added `symbol_fetcher.py` to handle API interactions for fetching symbols.
- Updated `downloader.py` to integrate this new symbol fetching logic.
- Added comprehensive unit tests for the new feature, including mocking
  API responses and error conditions.

All existing and new tests pass, ensuring functionality remains robust after these changes.
